refactor(tangle-dapp): Move wallet and network connection to shared ui lib
