+++ apps/tangle-dapp/components/ValidatorTable/ValidatorTable.tsx
@@ -11,6 +11,7 @@ import {
 } from '@tanstack/react-table';
 import { makeExplorerUrl } from '@webb-tools/api-provider-environment/transaction/utils';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
+import { ExplorerType } from '@webb-tools/tangle-shared-ui/types';
 import {
   Avatar,
   Button,
@@ -25,7 +26,7 @@ import Link from 'next/link';
 import { FC, useMemo, useState } from 'react';
 
 import { IS_PRODUCTION_ENV } from '../../constants/env';
-import { ExplorerType, PagePath, Validator } from '../../types';
+import { PagePath, Validator } from '../../types';
 import calculateCommission from '../../utils/calculateCommission';
 import {
   getSortAddressOrIdentityFnc,
