--- apps/tangle-dapp/components/WalletDropdown/index.ts
@@ -1,1 +1,0 @@
-export * from './WalletDropdown';
