+++ apps/tangle-dapp/containers/RecentTxContainer/RecentTxContainer.tsx
@@ -3,6 +3,7 @@
 import { makeExplorerUrl } from '@webb-tools/api-provider-environment/transaction/utils';
 import { ExternalLinkLine } from '@webb-tools/icons';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
+import { ExplorerType } from '@webb-tools/tangle-shared-ui/types';
 import { Button, Typography } from '@webb-tools/webb-ui-components';
 import { TANGLE_DOCS_URL } from '@webb-tools/webb-ui-components/constants';
 import { FC, useMemo } from 'react';
@@ -10,7 +11,6 @@ import { FC, useMemo } from 'react';
 import GlassCard from '../../components/GlassCard/GlassCard';
 import useActiveAccountAddress from '../../hooks/useActiveAccountAddress';
 import useAgnosticAccountInfo from '../../hooks/useAgnosticAccountInfo';
-import { ExplorerType } from '../../types';
 
 const RecentTxContainer: FC = () => {
   const { network } = useNetworkStore();
