+++ libs/tangle-shared-ui/src/components/ConnectWalletButton/index.ts
@@ -1,0 +1,3 @@
+import ConnectWalletButton from './ConnectWalletButton';
+
+export default ConnectWalletButton;
